+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Component `AffiliateProfile` added
+- Structure adjustments for the `Totalizer`, `Table` and `Topbar` components
+- Table measure fix
+
 ## [0.51.0] - 2022-07-07
 
 ### Added